% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md5sum.R
\name{md5sum}
\alias{md5sum}
\alias{compare_with_remote}
\title{calculation of multiple files in a directory}
\usage{
md5sum(
  path,
  pattern = "*.jpg",
  threads = round(parallel::detectCores() * 0.8),
  user = Sys.getenv("USER")
)

compare_with_remote(path, pattern = "*.jpg", zipfile = "files.zip")
}
\arguments{
\item{path}{path to the directory}

\item{pattern}{file pattern}

\item{threads}{number of core used}
}
\value{
a tibble
}
\description{
calculation of multiple files in a directory
}
